feat(streamserver): Finish the implementation of RTMPS server support

+++ pkg/streamserver/implementations/mediamtx/streamserver/rtmp_server.go
@@ -1,17 +1,23 @@
 package streamserver
 
 import (
-	"bytes"
 	"context"
 	"crypto"
+	"crypto/ed25519"
+	"crypto/rand"
 	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
 	"fmt"
-	"io"
+	"log"
+	"math/big"
 	"os"
+	"time"
 
+	"github.com/facebookincubator/go-belt/tool/logger"
 	"github.com/hashicorp/go-multierror"
 	"github.com/xaionaro-go/mediamtx/pkg/conf"
-	"github.com/xaionaro-go/mediamtx/pkg/logger"
+	mediamtxlogger "github.com/xaionaro-go/mediamtx/pkg/logger"
 	"github.com/xaionaro-go/mediamtx/pkg/pathmanager"
 	"github.com/xaionaro-go/mediamtx/pkg/servers/rtmp"
 	"github.com/xaionaro-go/streamctl/pkg/streamserver/types"
@@ -20,32 +26,38 @@ import (
 )
 
 type RTMPServer struct {
-	Locker xsync.Mutex
 	*rtmp.Server
+
+	locker         xsync.Mutex
+	originalConfig types.ServerConfig
+	isInitialized  bool
 }
 
 func newRTMPServer(
 	pathManager *pathmanager.PathManager,
 	listenAddr string,
-	logger logger.Writer,
+	logger mediamtxlogger.Writer,
 	opts ...types.ServerOption,
 ) (*RTMPServer, error) {
 	cfg := types.ServerOptions(opts).Config(context.Background())
-	srv := &RTMPServer{Server: &rtmp.Server{
-		Address:             listenAddr,
-		ReadTimeout:         conf.StringDuration(cfg.ReadTimeout),
-		WriteTimeout:        conf.StringDuration(cfg.WriteTimeout),
-		IsTLS:               cfg.IsTLS,
-		ServerCert:          "",
-		ServerKey:           "",
-		RTSPAddress:         "",
-		RunOnConnect:        "",
-		RunOnConnectRestart: false,
-		RunOnDisconnect:     "",
-		ExternalCmdPool:     nil,
-		PathManager:         pathManager,
-		Parent:              logger,
-	}}
+	srv := &RTMPServer{
+		originalConfig: cfg,
+		Server: &rtmp.Server{
+			Address:             listenAddr,
+			ReadTimeout:         conf.StringDuration(cfg.ReadTimeout),
+			WriteTimeout:        conf.StringDuration(cfg.WriteTimeout),
+			IsTLS:               cfg.IsTLS,
+			ServerCert:          "",
+			ServerKey:           "",
+			RTSPAddress:         "",
+			RunOnConnect:        "",
+			RunOnConnectRestart: false,
+			RunOnDisconnect:     "",
+			ExternalCmdPool:     nil,
+			PathManager:         pathManager,
+			Parent:              logger,
+		},
+	}
 	if err := srv.init(cfg); err != nil {
 		return nil, err
 	}
@@ -54,7 +66,12 @@ func newRTMPServer(
 
 func (srv *RTMPServer) init(
 	cfg types.ServerConfig,
-) error {
+) (_err error) {
+	defer func() {
+		if _err != nil {
+			_ = srv.Close()
+		}
+	}()
 	if (cfg.ServerCert != nil) != (cfg.ServerKey != nil) {
 		return fmt.Errorf("fields 'ServerCert' and 'ServerKey' should be used together (cur values: ServerCert == %#+v; ServerKey == %#+v)", cfg.ServerCert, cfg.ServerKey)
 	}
@@ -65,9 +82,63 @@ func (srv *RTMPServer) init(
 			return fmt.Errorf("unable to set the TLS certificate: %w", err)
 		}
 	}
+	if cfg.IsTLS && (srv.ServerCert == "" || srv.ServerKey == "") {
+		ctx := context.TODO()
+		logger.Warnf(ctx, "TLS is enabled, but no certificate is supplied, generating an ephemeral self-signed certificate") // TODO: implement the support of providing the certificates in the UI
+		err := srv.generateServerCertificate()
+		if err != nil {
+			return fmt.Errorf("unable to set the TLS certificate: %w", err)
+		}
+	}
 
 	if err := srv.Initialize(); err != nil {
 		return fmt.Errorf("Initialize() returned an error: %w", err)
+	}
+	srv.isInitialized = true
+
+	return nil
+}
+
+// Note! It create temporary files that should be cleaned up afterwards.
+func (srv *RTMPServer) generateServerCertificate() (_err error) {
+	publicKey, privateKey, err := ed25519.GenerateKey(rand.Reader)
+	if err != nil {
+		return fmt.Errorf("unable to generate an ED25519 keypair: %w", err)
+	}
+
+	serialNumber, err := rand.Int(rand.Reader, big.NewInt(1<<62))
+	if err != nil {
+		log.Fatalf("Error generating serial number: %v", err)
+	}
+
+	template := &x509.Certificate{
+		SerialNumber: serialNumber,
+		Subject: pkix.Name{
+			Organization: []string{"StreamPanel"},
+		},
+		NotBefore:             time.Date(1900, 1, 1, 0, 0, 0, 0, time.UTC),
+		NotAfter:              time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC),
+		KeyUsage:              x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+	}
+	certBytes, err := x509.CreateCertificate(rand.Reader, template, template, publicKey, privateKey)
+	if err != nil {
+		return fmt.Errorf("unable to generate a certificate: %w", err)
+	}
+
+	certificate, err := x509.ParseCertificate(certBytes)
+	if err != nil {
+		return fmt.Errorf("unable to parse the certificate that I've just generated: %w", err)
+	}
+
+	if certificate.PublicKey == nil {
+		return fmt.Errorf("internal error (bug in the code): public key is not set")
+	}
+
+	err = srv.setServerCertificate(*certificate, privateKey)
+	if err != nil {
+		return fmt.Errorf("unable to set the TLS certificate to an ephemeral one: %w", err)
 	}
 
 	return nil
@@ -97,9 +168,12 @@ func (srv *RTMPServer) setServerCertificate(
 	}
 	defer certFile.Close()
 
-	_, err = io.Copy(certFile, bytes.NewReader(cert.Raw))
-	if err != nil {
-		return fmt.Errorf("unable to write the server certificate to file '%s': %w", certFile.Name())
+	certPEM := &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: cert.Raw,
+	}
+	if err := pem.Encode(certFile, certPEM); err != nil {
+		return fmt.Errorf("unable to write the server certificate to file '%s' in PEM format: %w", certFile.Name(), err)
 	}
 
 	keyFile, err = os.CreateTemp("", "rtmps-server-certkey-*.pem")
@@ -108,14 +182,17 @@ func (srv *RTMPServer) setServerCertificate(
 	}
 	defer keyFile.Close()
 
-	keyBytes, err := x509.MarshalPKCS8PrivateKey(keyFile)
+	keyBytes, err := x509.MarshalPKCS8PrivateKey(key)
 	if err != nil {
 		return fmt.Errorf("unable to serialize into PKCS8 the private key: %w", err)
 	}
 
-	_, err = io.Copy(keyFile, bytes.NewReader(keyBytes))
-	if err != nil {
-		return fmt.Errorf("unable to write the server certificate to file '%s': %w", certFile.Name())
+	privatePem := &pem.Block{
+		Type:  "PRIVATE KEY",
+		Bytes: keyBytes,
+	}
+	if err := pem.Encode(keyFile, privatePem); err != nil {
+		return fmt.Errorf("unable to write the server certificate to file '%s' in PEM format: %w", certFile.Name(), err)
 	}
 
 	srv.ServerCert = certFile.Name()
@@ -125,21 +202,31 @@ func (srv *RTMPServer) setServerCertificate(
 
 var _ types.PortServer = (*RTMPServer)(nil)
 
+func (srv *RTMPServer) Config() types.ServerConfig {
+	return srv.originalConfig
+}
+
 func (srv *RTMPServer) Close() error {
 	ctx := context.TODO()
-	return xsync.DoR1(ctx, &srv.Locker, func() error {
+	return xsync.DoR1(ctx, &srv.locker, func() error {
 		if srv.Server == nil {
 			return fmt.Errorf("already closed")
 		}
 		certFile := srv.ServerCert
 		keyFile := srv.ServerKey
 
-		srv.Server.Close()
+		if srv.isInitialized {
+			srv.Server.Close()
+		}
 		srv.Server = nil
 
 		var result *multierror.Error
-		result = multierror.Append(result, os.Remove(certFile))
-		result = multierror.Append(result, os.Remove(keyFile))
+		if certFile != "" {
+			result = multierror.Append(result, os.Remove(certFile))
+		}
+		if keyFile != "" {
+			result = multierror.Append(result, os.Remove(keyFile))
+		}
 		return result.ErrorOrNil()
 	})
 }
